Tutorials/langfuse/requirements.txt
langfuse>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.38
python-dotenv>=1.0.0
